Only use nginx ssl proxy if mdata variable nginx_ssl is set
This allow you to run jenkins without an ssl proxy on the local machine. It's still recommended to use ssl, but maybe you've an proxy somewhere else. The Jenkins SMF script is automatically configured to listen on localhost or on all IP addresses. For that reason the SMF script has been modified to use an default instance. This allow us to configure it via svccfg properties.